refactor: Move issue and PR templates to global
Having issue and pull request templates in .github/.github allows them
to be available for all repositories if a similar PR or issue template
does not exist in that repository. So, these will not be used in
NOAA-FIMS/FIMS until the current default ones are removed but they will
be available in the other repositories.

This change partitions chores from the code base, which is nice for
linear commit histories and code protection rules.